PIIX: add disable-piix option in configure
Add disable-piix/enable-piix option, which can
disable/enable piix during compile. The PC platform
default is disabled.
