+++ configure
@@ -262,6 +262,7 @@ vhost_scsi="no"
 vhost_vsock="no"
 vhost_user=""
 kvm="no"
+piix="no"
 gprof="no"
 debug="no"
 sanitizers="no"
@@ -661,6 +662,10 @@ for opt do
   ;;
   --enable-kvm) kvm="yes"
   ;;
+  --disable-piix) piix="no"
+  ;;
+  --enable-piix) piix="yes"
+  ;;
   --disable-cap-ng)  cap_ng="no"
   ;;
   --enable-cap-ng) cap_ng="yes"
@@ -967,6 +972,7 @@ disabled with --disable-FEATURE, default is enabled if available:
   membarrier      membarrier system call (for Linux 4.14+ or Windows)
   fdt             fdt device tree
   kvm             KVM acceleration support
+  piix            support I440FX platform in kvm for x86
   vde             support for vde network
   netmap          support for netmap network
   linux-aio       Linux AIO support
@@ -3474,7 +3480,7 @@ echo "Linux AIO support $linux_aio"
 echo "ATTR/XATTR support $attr"
 echo "Install blobs     $blobs"
 echo "KVM support       $kvm"
-echo "TCG support       $tcg"
+echo "I440FX support    $piix"
 echo "malloc trim support $malloc_trim"
 echo "fdt support       $fdt"
 echo "membarrier        $membarrier"
@@ -4166,6 +4172,9 @@ if supported_kvm_target $target; then
         if test "$vhost_user" = "yes" ; then
             echo "CONFIG_VHOST_USER_NET_TEST_$target_name=y" >> $config_host_mak
         fi
+    fi
+    if test "$piix" = "yes" && test "$target_name" = "x86_64"; then
+        echo "CONFIG_PIIX=y" >> $config_host_mak
     fi
 fi
 if test "$target_bigendian" = "yes" ; then
