+++ default-configs/x86_64-softmmu.mak
@@ -19,7 +19,6 @@ CONFIG_MEM_HOTPLUG=y
 CONFIG_NVDIMM=y
 CONFIG_PCI=y
 CONFIG_PCIE_PORT=y
-CONFIG_PCI_PIIX=y
 CONFIG_PCI_Q35=y
 CONFIG_PFLASH_CFI01=y
 CONFIG_SERIAL=y
