+++ hw/i386/Makefile.objs
@@ -1,7 +1,9 @@
 obj-$(CONFIG_KVM) += kvm/
-obj-y += pc.o pc_piix.o pc_q35.o
+obj-y += pc.o pc_q35.o
+obj-y += acpi-build-core.o acpi-build-q35.o
+obj-$(CONFIG_PIIX) += pc_piix.o acpi-build-piix.o
 obj-y += pc_sysfw.o
 obj-$(CONFIG_VTD) += x86-iommu.o intel_iommu.o
 obj-$(CONFIG_AMD_IOMMU) += x86-iommu.o amd_iommu.o
 
-obj-y += acpi-build.o
+obj-$(call lnot,$(CONFIG_PIIX)) += acpi-build-stub.o
