+++ hw/i386/pc_piix.c
@@ -27,6 +27,7 @@
 #include "hw/hw.h"
 #include "hw/loader.h"
 #include "hw/i386/pc.h"
+#include "hw/i386/pc_piix.h"
 #include "hw/i386/apic.h"
 #include "hw/smbios/smbios.h"
 #include "hw/pci/pci.h"
