+++ stubs/Makefile.objs
@@ -35,5 +35,5 @@ stub-obj-y += target-monitor-defs.o
 stub-obj-y += target-get-monitor-def.o
 stub-obj-y += pc_madt_cpu_entry.o
 stub-obj-y += vmgenid.o
-stub-obj-y += pci-host-piix.o
+stub-obj-$(CONFIG_PIIX) += pci-host-piix.o
 stub-obj-y += ram-block.o
