+++ stubs/pci-host-piix.c
@@ -1,5 +1,6 @@
 #include "qemu/osdep.h"
-#include "hw/i386/pc.h"
+#include "hw/i386/pc_piix.h"
+
 PCIBus *find_i440fx(void)
 {
     return NULL;
